requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
bta-lib==1.0.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
dateparser==1.0.0
idna==3.2
multidict==5.1.0
numpy==1.21.2
pandas==1.3.3
pandas-ta==0.3.14b0
python-binance==1.0.14
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
tornado==6.1
typing-extensions==3.10.0.2
tzlocal==3.0
ujson==4.2.0
urllib3==1.26.7
websockets==10.0
yarl==1.6.3
